+++ jwst/outlier_detection/utils.py
@@ -36,16 +36,15 @@ def nanmedian3D(cube, overwrite_input=True):
     that only changes the dtype of the output but not the internal precision of nanmedian.
     It appears to be impossible to stop nanmedian from computing with 
     at least 64-bit precision internally.
-    Pre-allocating the output array, i.e., out=np.empty(cube.shape[1:], dtype=np.float32)
+    Pre-allocating the output array with `out=np.empty(cube.shape[1:], dtype=np.float32)`
     actually increases memory usage a bit, so it's better to just convert after the fact.
-    Returning float64 here causes gwcs_blot to segfault down the line 
+    Failing to return float32 here causes gwcs_blot to segfault down the line 
     because that function is hard-coded to expect float32. 
     """
     with warnings.catch_warnings():
         warnings.filterwarnings(action="ignore",
                                 message="All-NaN slice encountered",
                                 category=RuntimeWarning)
-
         return np.nanmedian(cube, axis=0, overwrite_input=overwrite_input).astype(np.float32)
 
 
@@ -63,30 +62,51 @@ def create_cube_median(cube_model, maskpt):
 def drizzle_and_median(input_models,
                        resamp,
                        maskpt,
-                       make_output_path=None,
                        resample_data=False,
-                       in_memory=True,
                        save_intermediate_results=False,
+                       make_output_path=None,
                        buffer_size=None):
     """
+    Shared code between imaging and spec modes for resampling and median computation
+
     Parameters
     ----------
     input_models : ModelLibrary
-        The input datamodels
+        The input datamodels.
 
-    resamp : ResampleData object
+    resamp : resample.resample.ResampleData object
+        The controlling object for the resampling process.
 
     maskpt : float
+        The weight threshold for masking out low weight pixels.
 
-    make_output_path : ?
+    resample_data : bool
+        Whether or not to do resampling.
+
+    save_intermediate_results : bool
+        if True, save the drizzled models and median model to fits.
+
+    make_output_path : function
+        The functools.partial instance to pass to save_median. Must be 
+        specified if save_intermediate_results is True. Default None.
+
+    buffer_size : int
+        The size of chunk in bytes that will be read into memory when computing the median.
+        This parameter has no effect if the input library has its on_disk attribute
+        set to False.
     """
+
+    # validate inputs
     if save_intermediate_results and (make_output_path is None):
         raise ValueError("make_output_path is required if save_intermediate_results is True")
     if not resamp.single:
         raise ValueError("drizzle_and_median should only be used for resample_many_to_many")
+    in_memory = not input_models._on_disk
+
     if resample_data:
         indices_by_group = list(input_models.group_indices.values())
     else:
+        # treat each input model as if it is the only member of its group
         indices_by_group = [[i] for i in range(len(input_models))]
 
     with input_models:
@@ -94,9 +114,7 @@ def drizzle_and_median(input_models,
 
             if resample_data:
                 median_wcs = resamp.output_wcs
-
                 drizzled_model = resamp.resample_group(input_models, indices)
-
             else:
                 # for non-dithered data, the resampled image is just the original image
                 drizzled_model = input_models.borrow(i)
@@ -105,15 +123,17 @@ def drizzle_and_median(input_models,
                     weight_type=resamp.weight_type,
                     good_bits=resamp.good_bits)
                 input_models.shelve(drizzled_model, i, modify=True)
+                # copy for when saving median and input is a filename?
                 median_wcs = copy.deepcopy(drizzled_model.meta.wcs)
 
             if i == 0:
-                # allocate either temporary storage or memory for data arrays that go into median
                 ngroups = len(indices_by_group)
                 full_shape = (ngroups,) + drizzled_model.data.shape
                 if in_memory:
+                    # allocate memory for data arrays that go into median
                     data_frames = np.empty(full_shape, dtype=np.float32)
                 else:
+                    # set up temporary storage for data arrays that go into median
                     median_computer = OnDiskMedian(full_shape,
                                                     dtype=drizzled_model.data.dtype,
                                                     buffer_size=buffer_size)
@@ -130,10 +150,11 @@ def drizzle_and_median(input_models,
             weight_threshold = compute_weight_threshold(drizzled_model.wht, maskpt)
             drizzled_model.data[drizzled_model.wht < weight_threshold] = np.nan
 
-            # put drizzled models into either storage or memory for median
             if in_memory:
+                # populate pre-allocated memory with the drizzled data
                 data_frames[i] = drizzled_model.data
             else:
+                # distribute the drizzled data into the temporary storage
                 median_computer.add_image(drizzled_model.data)
 
     # Perform median combination on set of drizzled mosaics
@@ -145,7 +166,7 @@ def drizzle_and_median(input_models,
         median_computer.cleanup()
 
     if save_intermediate_results:
-        # make a median model
+        # Save median model to fits
         median_model = datamodels.ImageModel(median_data)
         median_model.update(drizzled_model)
         median_model.meta.wcs = median_wcs
@@ -221,7 +242,7 @@ class DiskAppendableArray:
     
 
     def cleanup(self):
-        Path.unlink(self._filename)
+        Path.unlink(Path(self._filename))
         return
 
 
@@ -287,16 +308,16 @@ class OnDiskMedian:
         imrows, imcols = self.frame_shape
         if buffer_size is None:
             buffer_size = imrows * imcols * self.itemsize
-        self.buffer_size = buffer_size
         per_model_buffer_size = buffer_size / self._expected_nframes
         min_buffer_size = imcols * self.itemsize
         section_nrows = min(imrows, int(per_model_buffer_size // min_buffer_size))
 
-        if section_nrows == 0:
+        if section_nrows <= 0:
             buffer_size = min_buffer_size * self._expected_nframes
             log.warning("Buffer size is too small to hold a single row."
                             f"Increasing buffer size to {buffer_size / _ONE_MB}MB")
             section_nrows = 1
+        self.buffer_size = buffer_size
 
         nsections = int(np.ceil(imrows / section_nrows))
         log.info(f"Computing median over {self._expected_nframes} images in {nsections} "
@@ -340,7 +361,7 @@ class OnDiskMedian:
     def cleanup(self):
         """Remove the temporary files and directory when finished"""
         [arr.cleanup() for arr in self._temp_arrays]
-        Path.rmdir(self._temp_path)
+        self._temp_dir.cleanup()
         return